refactor and updated some code
