+++ feed/index.html
@@ -323,7 +323,7 @@
       <div class="hidden lg:block">
         <div
           id="aside-feed-search-container"
-          class="hidden flex justify-center transition-all max-w-xl mx-auto mt-2 px-4 py-1 rounded-lg focus-within:bg-black focus-within:ring ring-eden-400 focus-within:bg-transparent">
+          class="hidden flex-col justify-center transition-all max-w-xl mx-auto mt-2 px-4 py-1 rounded-lg focus-within:bg-black focus-within:ring ring-eden-400 focus-within:bg-transparent">
           <input
             type="search"
             id="aside-feed-search"
